feat: refactor chat components by replacing ChatWindowView and ChatSuggestView with ChatWindow and ChatSuggest for improved structure and functionality
